zoomTo         # 1-second smooth zoom in (factor = 2) zoomTo out     # 1-second smooth zoom out (factor = 1/2) zoomTo (atom)  #1 second smooth pan to this atom if off it zoomTo (atom)  #1 second smooth zoom in at this atom if at it
zoomTo [time-sec] (atom)
zoomTo [time-sec] (atom) zoom
zoomTo [time-sec] (atom) +delta|-delta
zoomTo [time-sec] (atom) *factor|/factor

zoom is similar, but simpler with the following differences

zoom     # enable zoom
zoom ON  # enable zoom
zoom OFF # disable zoom

no automatic zooming in and out 

note that "set picking center" now uses "zoomTo" exclusively,
and actuates "zoomTo out" if OFF the structure

+++ Jmol/src/org/jmol/viewer/MouseManager.java
@@ -428,7 +428,7 @@ abstract class MouseManager {
         while (++rotation <= 0)
           zoomLevel *= wheelClickFractionDown;
       }
-      viewer.zoomToPercent((int)(zoomLevel * 100 + 0.5f));
+      viewer.zoomToPercent(zoomLevel * 100 + 0.5f);
     }
   }
